Added ability to call function without parenthesis
Calling functions is hence similar to Haskell now. Since function
application binds strongest, the following

func a * b

is parsed as

(func a) * b

and needs to be written as

func (a*b)

if func is to be applied to the product of a and b.
Also, it follows that

func func a  == (func func) a

is illegal.